+++ app/schemas/purchase.py
@@ -4,16 +4,16 @@ from datetime import datetime
 
 class PurchaseRequest(BaseModel):
     clientId: int = Field(..., examples=[1])
-    amount: Decimal = Field(..., gt=0, examples=[250])
+    amount: float = Field(..., gt=0, examples=[250])
     currency: str = Field(..., min_length=3, max_length=3, examples=["USD"])
     purchaseDate: datetime = Field(..., examples=["2025-09-20T14:30:00Z"])
     purchaseCountry: str = Field(..., examples=["Francia"])
 
 class PurchaseSummary(BaseModel):
     clientId: int
-    originalAmount: Decimal
-    discountApplied: Decimal
-    finalAmount: Decimal
+    originalAmount: float
+    discountApplied: float
+    finalAmount: float
     benefit: str | None = None
 
 class PurchaseOK(BaseModel):
